🐛 fix: split Header components with Logo, Header Description, Header Features
