+++ src/shared/ui/Header/HeaderDescription/index.ts
@@ -1,0 +1,1 @@
+export { default as HeaderDescription } from "./HeaderDescription";
