+++ src/shared/ui/Header/HeaderFeatures/index.ts
@@ -1,0 +1,1 @@
+export { default as HeaderFeatures } from "./HeaderFeatures";
