+++ src/shared/ui/Header/index.ts
@@ -1,1 +1,4 @@
 export { default as Header } from "./Header";
+export * from "./Logo";
+export * from "./HeaderDescription";
+export * from "./HeaderFeatures";
